move pyVimeo code file to inner folder
