--- backend/engine/uploader/Vimeo/tusclient/uploader/__init__.py
@@ -1,1 +1,0 @@
-from engine.uploader.Vimeo.tusclient.uploader.uploader import Uploader
